separate sequencer from main
