+++ trollup-sequencer/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod node;
+pub mod sequencer;
 pub mod server;
